Add support for multiplier constraints derived from axis anchors
The NSLayoutConstraint API allows for creation of constraints with both
offset and multiplier not only between dimensions (width & height), but also
between axial attributes (left, right, leading, trailing, center). Such
constraints are useful for relative positioning of a subview within a parent
view (e.g. position subview so that it's center point is located in the middle
of its parent view on horizontal axis and at 25% of the parent view's height
on the vertical axis).

For some reason, Apple chose not to support creation of such constraints through
the base NSLayoutAnchor API, and multiplier constraints are only exposed through the
NSLayoutDimension subclass. Since Anchorage is wrapping the NSLayoutAnchor API, it
also lacks these axial multiplier constraints.

This commit introduces methods for creating multiplier constraints between axis anchors
in extensions on NSLayoutXAxisAnchor and NSLayoutYAxisAnchor, as well as corresponding
new functions to the Anchorage API.